command/add: use a bool for FromState, instead of an address
After some internal back and forth, we've settled on `-from-state` as a bool, which limits users to generating configuration from the same address as the target address (which is to say: generate config for something that's already been imported). We decided that it would be easier to add something like a "-force" flag to allow generating config from a different resource than it would be to change the flag from a string to a bool after the fact.